+++ src/serde/mod.rs
@@ -1,0 +1,6 @@
+mod deser_skin;
+mod deser_compound_style;
+mod deser_line_style;
+mod deser_styled_char;
+
+